refactor(phase-5d): split developer projects actions into cohesive modules
Break the 729-line actions.ts into actions/{package,screenshot,release,
pricing}.ts (each "use server", async actions only) plus a non-use-server
actions/_shared.ts holding the shared helpers (sameUser, createDedicatedFile),
zod schemas, and State/ReleaseRecord types. Update the 8 consumers to import
from the new modules; delete the old actions.ts (no barrel).

All 15 server actions moved verbatim; the use-server constraint (only async
exports) is respected; no behavior change. lint + tsc green.
