--- src/app/[lang]/(developer)/developer/projects/[name]/actions.ts
@@ -1,734 +1,0 @@
-"use server";
-
-import { addAuditLog, auditLogActions } from "@/lib/audit-log";
-import type { ActionResult } from "@/lib/action-result";
-import { authenticated, throwIfUnauth } from "@/lib/auth-guard";
-import { isAdmin } from "@/lib/admin-guard";
-import { contentPath } from "@/lib/content-url";
-import {
-  createDevPackageScreenshot,
-  deleteDevPackage,
-  deleteDevPackageScreenshot,
-  getPackageNameFromPackageId,
-  getPackagePublishedByIdOrThrow,
-  getUserIdFromPackageId,
-  retrieveDevPackageByName,
-  retrieveDevPackageDependsFile,
-  retrieveDevPackageIconFile,
-  retrieveDevPackageLastScreenshotOrder,
-  retrieveDevPackageScreenshots,
-  updateDevPackageDescription,
-  updateDevPackageDisplay,
-  updateDevPackageIconFile,
-  updateDevPackagePublished,
-  updateDevPackageScreenshotOrder,
-  updateDevPackageTags,
-  updatePackageInterval,
-  upsertPackagePricings,
-} from "@/lib/db/package";
-import {
-  createRelease as createReleaseRecord,
-  deleteReleaseById,
-  getReleasePackageAndFileId,
-  getReleasePublishedByIdOrThrow,
-  getReleaseWithFileById,
-  updateRelease as updateReleaseRecord,
-} from "@/lib/db/release";
-import type { PaymentInterval } from "@prisma/client";
-import { isValidNuGetVersionRange } from "@/lib/nuget-version-range";
-import {
-  calcTotalFileSize,
-  createStorageFile,
-  deleteStorageFile,
-} from "@/lib/storage";
-import { getLanguage } from "@/lib/lang-utils";
-import { revalidatePath } from "next/cache";
-import { redirect } from "next/navigation";
-import SemVer from "semver";
-import { z } from "zod";
-
-export type State = {
-  errors?: {
-    displayName?: string[];
-    shortDescription?: string[];
-  };
-  success?: boolean;
-  message?: string | null;
-};
-type ReleaseRecord = Awaited<ReturnType<typeof updateReleaseRecord>>;
-
-const displayNameAndShortDescriptionSchema = z.object({
-  displayName: z.string().max(50, "表示名は50文字以下である必要があります"),
-  shortDescription: z
-    .string()
-    .max(200, "短い説明は200文字以下である必要があります"),
-  id: z.string().uuid("IDが不正です"),
-});
-
-const releaseSchema = z.object({
-  title: z.string().max(50, "タイトルは50文字以下である必要があります"),
-  description: z.string().max(1000, "説明は1000文字以下である必要があります"),
-  id: z.string().uuid("IDが不正です"),
-  targetVersion: z
-    .string()
-    .refine(
-      (v) => SemVer.valid(v) || isValidNuGetVersionRange(v),
-      "バージョンが不正です",
-    ),
-  published: z
-    .string()
-    .refine((v) => v === "on" || v === "off", "公開設定が不正です"),
-  file: z.any()
-});
-
-async function sameUser<TResult>(
-  packageId: string,
-  userId: string,
-  fnc: () => Promise<TResult>,
-) {
-  const pkgUserId = await getUserIdFromPackageId({ packageId });
-  if (!pkgUserId) {
-    return {
-      success: false,
-      message: "IDが見つかりません",
-    };
-  }
-
-  if (pkgUserId !== userId) {
-    return {
-      success: false,
-      message: "権限がありません",
-    };
-  }
-
-  return await fnc();
-}
-
-async function createDedicatedFile(userId: string, file: File, size: bigint) {
-  const maxSize = BigInt(1024 * 1024 * 1024); // 1GB
-  let totalSize = await calcTotalFileSize({ userId });
-  totalSize -= size;
-
-  if (totalSize + BigInt(file.size) > maxSize) {
-    return {
-      success: false,
-      message: "ストレージ容量が足りません",
-    };
-  }
-
-  const record = await createStorageFile({
-    file,
-    visibility: "DEDICATED",
-    userId,
-  });
-
-  return {
-    success: true,
-    record,
-  };
-}
-
-export async function updateDisplayNameAndShortDescription(
-  state: State,
-  formData: FormData,
-): Promise<State> {
-  return await authenticated(async (session) => {
-    const validated = displayNameAndShortDescriptionSchema.safeParse(
-      Object.fromEntries(formData.entries()),
-    );
-    if (!validated.success) {
-      return {
-        errors: validated.error.flatten().fieldErrors,
-        message: "入力内容に誤りがあります",
-        success: false,
-      };
-    }
-
-    const { displayName, shortDescription, id } = validated.data;
-    return await sameUser(id, session.user.id, async () => {
-      const { name } = await updateDevPackageDisplay({
-        packageId: id,
-        displayName,
-        shortDescription,
-      });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.updatePackage,
-        details: `packageId: ${id}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-      };
-    });
-  });
-}
-
-export async function updateDescription({
-  packageId,
-  description,
-}: { packageId: string; description: string }): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    if (description.length > 1000) {
-      return {
-        message: "説明は1000文字以下である必要があります",
-        success: false,
-      };
-    }
-
-    return await sameUser(packageId, session.user.id, async () => {
-      const { name } = await updateDevPackageDescription({
-        packageId,
-        description,
-      });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.updatePackage,
-        details: `packageId: ${packageId}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-      };
-    });
-  });
-}
-
-export async function retrievePackage(name: string) {
-  const session = await throwIfUnauth();
-  const pkg = await retrieveDevPackageByName({ name, userId: session.user.id });
-  if (!pkg) {
-    return null;
-  }
-  pkg.Release.sort((a, b) => {
-    return new SemVer.SemVer(b.version).compare(a.version);
-  });
-  const screenshots = await Promise.all(
-    pkg.PackageScreenshot.map(async (item) => {
-      return {
-        ...item,
-        url: contentPath(item.file.id),
-      };
-    }),
-  );
-
-  return {
-    ...pkg,
-    iconFileUrl: pkg.iconFile && contentPath(pkg.iconFile.id),
-    PackageScreenshot: screenshots,
-  };
-}
-
-export async function deletePackage(id: string): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    return await sameUser(id, session.user.id, async () => {
-      const files = await retrieveDevPackageDependsFile({ packageId: id });
-      await deleteDevPackage({ packageId: id });
-      const promises = files.map(async (file) => {
-        await deleteStorageFile({
-          fileId: file.id,
-        });
-      });
-
-      await Promise.all(promises);
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.deletePackage,
-        details: `packageId: ${id}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer`);
-      redirect(`/${lang}/developer`);
-    });
-  });
-}
-
-export async function changePackageVisibility(
-  id: string,
-  published: boolean,
-): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    return await sameUser(id, session.user.id, async () => {
-      const { published: oldPublished } = await getPackagePublishedByIdOrThrow({
-        id,
-      });
-      const { name } = await updateDevPackagePublished({
-        packageId: id,
-        published,
-      });
-      if (oldPublished !== published) {
-        await addAuditLog({
-          userId: session.user.id,
-          action: published
-            ? auditLogActions.developer.publishPackage
-            : auditLogActions.developer.unpublishPackage,
-          details: `packageId: ${id}`,
-        });
-      }
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-      };
-    });
-  });
-}
-
-export async function uploadIcon(formData: FormData): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    const file = formData.get("file") as File;
-    if (!file) {
-      return {
-        success: false,
-        message: "ファイルが見つかりません",
-      };
-    }
-    const id = formData.get("id") as string;
-    return await sameUser(id, session.user.id, async () => {
-      const iconFile = await retrieveDevPackageIconFile({ packageId: id });
-
-      const deletedSize = iconFile ? BigInt(iconFile.size) : BigInt(0);
-      const result = await createDedicatedFile(
-        session.user.id,
-        file,
-        deletedSize,
-      );
-      if (!result.success) {
-        return {
-          success: result.success,
-          message: result.message,
-        };
-      }
-
-      if (iconFile) {
-        await deleteStorageFile({
-          fileId: iconFile.id,
-        });
-      }
-
-      const { name } = await updateDevPackageIconFile({
-        packageId: id,
-        fileId: result.record!.id,
-      });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.updatePackage,
-        details: `packageId: ${id}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-      };
-    });
-  });
-}
-
-export async function addScreenshot(formData: FormData): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    const file = formData.get("file") as File;
-    if (!file) {
-      return {
-        success: false,
-        message: "ファイルが見つかりません",
-      };
-    }
-    const id = formData.get("id") as string;
-    return await sameUser(id, session.user.id, async () => {
-      const name = await getPackageNameFromPackageId({ packageId: id });
-      const result = await createDedicatedFile(
-        session.user.id,
-        file,
-        BigInt(0),
-      );
-      if (!result.success) {
-        return result;
-      }
-      const lastScreenshot = await retrieveDevPackageLastScreenshotOrder({
-        packageId: id,
-      });
-      await createDevPackageScreenshot({
-        packageId: id,
-        fileId: result.record!.id,
-        order: (lastScreenshot?.order || 0) + 1,
-      });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.updatePackage,
-        details: `packageId: ${id}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-      };
-    });
-  });
-}
-
-export async function moveScreenshot({
-  delta,
-  packageId,
-  fileId,
-}: { delta: number; packageId: string; fileId: string }): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    return await sameUser(packageId, session.user.id, async () => {
-      const name = await getPackageNameFromPackageId({ packageId });
-      const sign = Math.sign(delta);
-      const all = await retrieveDevPackageScreenshots({ packageId });
-
-      const target = all.find((screenshot) => screenshot.fileId === fileId);
-      if (!target) {
-        return {
-          success: false,
-          message: "ファイルが見つかりません",
-        };
-      }
-
-      const index = all.indexOf(target);
-      if (index === 0 && sign < 0) {
-        return {
-          success: false,
-          message: "既に先頭です",
-        };
-      }
-      if (index === all.length - 1 && sign > 0) {
-        return {
-          success: false,
-          message: "既に末尾です",
-        };
-      }
-      all.splice(index, 1);
-      all.splice(index + sign, 0, target);
-      const promises = all
-        .map((item, i) => ({
-          ...item,
-          order: i,
-          originalOrder: item.order,
-        }))
-        .map(async (item) => {
-          if (item.order === item.originalOrder) return;
-          await updateDevPackageScreenshotOrder({
-            packageId,
-            fileId: item.fileId,
-            order: item.order,
-          });
-        });
-      await Promise.all(promises);
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.updatePackage,
-        details: `packageId: ${packageId}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-      };
-    });
-  });
-}
-
-export async function deleteScreenshot({
-  packageId,
-  fileId,
-}: { packageId: string; fileId: string }): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    return await sameUser(packageId, session.user.id, async () => {
-      const name = await getPackageNameFromPackageId({ packageId });
-      await deleteDevPackageScreenshot({ packageId, fileId });
-      await deleteStorageFile({ fileId });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.updatePackage,
-        details: `packageId: ${packageId}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-      };
-    });
-  });
-}
-
-export async function updateTag({
-  packageId,
-  tags,
-}: { packageId: string; tags: string[] }): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    return await sameUser(packageId, session.user.id, async () => {
-      const { name } = await updateDevPackageTags({
-        packageId,
-        tags,
-      });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.updatePackage,
-        details: `packageId: ${packageId}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-      };
-    });
-  });
-}
-
-export async function updateRelease(
-  formData: FormData,
-): Promise<ActionResult<ReleaseRecord>> {
-  return await authenticated(async (session) => {
-    const validated = releaseSchema.safeParse(
-      Object.fromEntries(formData.entries()),
-    );
-    if (!validated.success) {
-      return {
-        errors: validated.error.flatten().fieldErrors,
-        message: "入力内容に誤りがあります",
-        success: false,
-      };
-    }
-    const release = await getReleaseWithFileById({
-      id: validated.data.id,
-    });
-    if (!release?.packageId) {
-      return {
-        success: false,
-        message: "IDが見つかりません",
-      };
-    }
-
-    return await sameUser(release.packageId, session.user.id, async () => {
-      let fileId = release.file?.id;
-      if (validated.data.file) {
-        const deletedSize = release.file
-          ? BigInt(release.file.size)
-          : BigInt(0);
-        const result = await createDedicatedFile(
-          session.user.id,
-          validated.data.file as File,
-          deletedSize,
-        );
-        if (!result.success) {
-          return {
-            success: result.success,
-            message: result.message,
-          };
-        }
-
-        if (release.file) {
-          await deleteStorageFile({
-            fileId: release.file.id,
-          });
-        }
-        fileId = result.record!.id;
-      }
-
-      const { published: oldPublished } = await getReleasePublishedByIdOrThrow({
-        id: validated.data.id,
-      });
-      const data = await updateReleaseRecord({
-        id: validated.data.id,
-        title: validated.data.title,
-        description: validated.data.description,
-        targetVersion: validated.data.targetVersion,
-        published: validated.data.published === "on",
-        fileId: fileId,
-      });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.updateRelease,
-        details: `releaseId: ${data.id}`,
-      });
-      if (oldPublished !== data.published) {
-        await addAuditLog({
-          userId: session.user.id,
-          action: data.published
-            ? auditLogActions.developer.publishRelease
-            : auditLogActions.developer.unpublishRelease,
-          details: `releaseId: ${data.id}`,
-        });
-      }
-
-      const lang = await getLanguage();
-      const name = await getPackageNameFromPackageId({ packageId: release.packageId });
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-
-      return {
-        success: true,
-        data,
-      };
-    });
-  });
-}
-
-export async function createRelease({
-  packageId,
-  version,
-}: {
-  packageId: string;
-  version: string;
-}): Promise<ActionResult<ReleaseRecord>> {
-  return await authenticated(async (session) => {
-    return await sameUser(packageId, session.user.id, async () => {
-      if (SemVer.valid(version) === null) {
-        return {
-          success: false,
-          message: "バージョンが不正です",
-        };
-      }
-
-      const release = await createReleaseRecord({
-        packageId,
-        version,
-        title: "新しいリリース",
-        description: "",
-        targetVersion: "1.0.0-preview.10",
-        published: false,
-      });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.createRelease,
-        details: `packageId: ${packageId}, releaseId: ${release.id}, version: ${version}`,
-      });
-      const lang = await getLanguage();
-      const name = await getPackageNameFromPackageId({ packageId });
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-        data: release,
-      };
-    });
-  });
-}
-
-export async function deleteRelease({
-  releaseId,
-}: { releaseId: string }): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    const release = await getReleasePackageAndFileId({
-      id: releaseId,
-    });
-    if (!release?.packageId) {
-      return {
-        success: false,
-        message: "IDが見つかりません",
-      };
-    }
-    return await sameUser(release.packageId, session.user.id, async () => {
-      if (release.fileId) {
-        await deleteStorageFile({
-          fileId: release.fileId,
-        });
-      }
-      await deleteReleaseById({
-        id: releaseId,
-      });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.developer.deleteRelease,
-        details: `releaseId: ${releaseId}`,
-      });
-      const lang = await getLanguage();
-      const name = await getPackageNameFromPackageId({
-        packageId: release.packageId,
-      });
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return {
-        success: true,
-      };
-    });
-  });
-}
-
-const pricingSchema = z.object({
-  packageId: z.string().uuid(),
-  pricings: z.array(
-    z.object({
-      currency: z.string().length(3, "通貨コードは3文字である必要があります"),
-      price: z.number().int().min(0, "価格は0以上である必要があります"),
-      fallback: z.boolean(),
-    }),
-  ).refine(
-    (pricings) => pricings.filter((p) => p.fallback).length <= 1,
-    "デフォルト通貨は1つまでです",
-  ),
-});
-
-export async function updatePricing({
-  packageId,
-  pricings,
-}: {
-  packageId: string;
-  pricings: { currency: string; price: number; fallback: boolean }[];
-}): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    if (!isAdmin(session.user.id)) {
-      return { success: false, message: "権限がありません" };
-    }
-
-    const validated = pricingSchema.safeParse({ packageId, pricings });
-    if (!validated.success) {
-      return {
-        success: false,
-        message: "入力内容に誤りがあります",
-      };
-    }
-
-    return await sameUser(packageId, session.user.id, async () => {
-      await upsertPackagePricings({ packageId, pricings });
-      const name = await getPackageNameFromPackageId({ packageId });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.admin.updatePackagePricing,
-        details: `packageId: ${packageId}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return { success: true };
-    });
-  });
-}
-
-const intervalSchema = z.object({
-  packageId: z.string().uuid(),
-  interval: z.enum(["ONCE", "MONTHLY", "YEARLY"]).nullable(),
-});
-
-export async function updateInterval({
-  packageId,
-  interval,
-}: {
-  packageId: string;
-  interval: PaymentInterval | null;
-}): Promise<ActionResult> {
-  return await authenticated(async (session) => {
-    if (!isAdmin(session.user.id)) {
-      return { success: false, message: "権限がありません" };
-    }
-
-    const validated = intervalSchema.safeParse({ packageId, interval });
-    if (!validated.success) {
-      return {
-        success: false,
-        message: "入力内容に誤りがあります",
-      };
-    }
-
-    return await sameUser(packageId, session.user.id, async () => {
-      const { name } = await updatePackageInterval({ packageId, interval });
-      await addAuditLog({
-        userId: session.user.id,
-        action: auditLogActions.admin.updatePackageInterval,
-        details: `packageId: ${packageId}, interval: ${interval}`,
-      });
-      const lang = await getLanguage();
-      revalidatePath(`/${lang}/developer/projects/${name}`);
-      return { success: true };
-    });
-  });
-}
