+++ src/app/[lang]/(developer)/developer/projects/[name]/types.ts
@@ -1,3 +1,3 @@
-import type { retrievePackage } from "./actions";
+import type { retrievePackage } from "./actions/package";
 
 export type Package = NonNullable<Awaited<ReturnType<typeof retrievePackage>>>;
